+++ au3/modules/import-export/RegisterImportPlugins.cpp
@@ -6,11 +6,17 @@
 #include "RegisterImportPlugins.h"
 
 #include "modules/import-export/mod-wavpack/ImportWavPack.h"
+#include "modules/import-export/mod-mpg123/ImportMP3_MPG123.h"
 
 void RegisterImportPlugins()
 {
     static Importer::RegisteredImportPlugin sWavPackImportPlugin{
         "WavPack",
         std::make_unique< WavPackImportPlugin >()
+    };
+
+    static Importer::RegisteredImportPlugin sMP3ImportPlugin{
+        "MP3",
+        std::make_unique<MP3ImportPlugin>()
     };
 }
