+++ buildscripts/cmake/SetupDependencies.cmake
@@ -67,6 +67,7 @@ populate(expat "expat/2.0.5")
 populate(portaudio "portaudio/19.7.0")
 populate(libmp3lame "libmp3lame/3.100")
 populate(wavpack "wavpack/5.7.0")
+populate(mpg123 "mpg123/1.31.2")
 
 if (NOT OS_IS_LIN)
     populate(zlib "zlib/1.2.13")
